CSSTUDIO-1827 Move "LinearTicksTest" from "org.csstudio.javafx.rtplot" to "org.csstudio.javafx.rtplot.internal", since the class "LinarTicks" that is being tested in that class is located in "org.csstudio.javafx.rtplot.internal". In the class "TicksTestBase", make the fields "buf" and "gc" "protected" in order to enable access to these fields from "LinearTicksTest" after the move.
